safari-534.53-branch/LayoutTests/fast/dom/Geolocation/disconnected-frame-permission-denied-expected.txt

frame "<!--framePath //<!--frame0-->-->" - has 1 onunload handler(s)
Tests that when a request is made on a Geolocation object, permission is denied and its Frame is disconnected before a callback is made, the error callback is invoked with PERMISSION_DENIED.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error.code is error.PERMISSION_DENIED
PASS error.message is "User denied Geolocation"

PASS error.code is error.PERMISSION_DENIED
PASS error.message is "User denied Geolocation"

PASS successfullyParsed is true

TEST COMPLETE
